test: 레거시 테스트 패키지 및 네이밍 정리
