+++ src/test/java/com/threestar/trainus/domain/coupon/user/concurrency/UserCouponDistributedLockIntegrationTest.java
@@ -1,6 +1,4 @@
-package com.threestar.trainus.coupon.user;
-
-import static org.assertj.core.api.AssertionsForClassTypes.*;
+package com.threestar.trainus.domain.coupon.user.concurrency;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
@@ -29,19 +27,15 @@ import com.threestar.trainus.domain.coupon.user.service.CouponService;
 import com.threestar.trainus.domain.user.entity.User;
 import com.threestar.trainus.domain.user.entity.UserRole;
 import com.threestar.trainus.domain.user.repository.UserRepository;
-import com.threestar.trainus.global.exception.handler.BusinessException;
 
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @SpringBootTest
-class UserCouponServiceConcurrencyTests {
+public class UserCouponDistributedLockIntegrationTest {
 
 	@Autowired
 	CouponIssueFacade couponIssueFacade;
-
-	@Autowired
-	CouponService couponService;
 
 	@Autowired
 	CouponRepository couponRepository;
@@ -76,7 +70,7 @@ class UserCouponServiceConcurrencyTests {
 		// 쿠폰 생성
 		coupon = couponRepository.save(
 			Coupon.builder()
-				.name("선착순 쿠폰")
+				.name("Redis 분산락 쿠폰")
 				.quantity(COUPON_QUANTITY)
 				.category(CouponCategory.OPEN_RUN)
 				.status(CouponStatus.ACTIVE)
@@ -91,14 +85,13 @@ class UserCouponServiceConcurrencyTests {
 
 	@AfterEach
 	void tearDown() {
-		// 테스트 후 데이터 정리
 		userCouponRepository.deleteAll();
 		couponRepository.deleteAll();
 		userRepository.deleteAll();
 	}
 
 	@Test
-	@DisplayName("동시 요청 시 - 비관적 락 적용: 발급된 쿠폰 수량만큼 발급")
+	@DisplayName("동시 요청 시 - Redis 분산락 적용: 발급된 쿠폰 수량만큼 발급")
 	void 쿠폰_동시_발급_테스트() throws InterruptedException {
 		ExecutorService executor = Executors.newFixedThreadPool(100);
 		CountDownLatch latch = new CountDownLatch(CONCURRENT_USERS);
@@ -114,14 +107,13 @@ class UserCouponServiceConcurrencyTests {
 
 			executor.submit(() -> {
 				try {
-
 					couponIssueFacade.issueCoupon(users.get(idx).getId(), coupon.getId());
-					log.info("발급 성공 - id: {} | 발급된 쿠폰 수량: {}", idx,
+					log.info("[성공] userId: {} | 현재 발급 수: {}", idx,
 						userCouponRepository.countByCouponId(coupon.getId()));
 					successCount.incrementAndGet();
 				} catch (Exception e) {
 					failCount.incrementAndGet();
-					log.error("발급 실패 - id: {} | message: {}", idx, e.getMessage());
+					log.error("[실패] userId: {} | message: {}", idx, e.getMessage());
 				} finally {
 					latch.countDown();
 				}
@@ -129,60 +121,20 @@ class UserCouponServiceConcurrencyTests {
 		}
 
 		latch.await();
-		stopWatch.stop();
 		executor.shutdown();
+		stopWatch.stop();
 
 		long issuedCount = userCouponRepository.countByCouponId(coupon.getId());
 		int remainingQuantity = couponRepository.findById(coupon.getId()).orElseThrow().getQuantity();
 
 		log.info("총 소요 시간(ms): {}", stopWatch.getTotalTimeMillis());
-		log.info("요청 총 수: {}", CONCURRENT_USERS);
+		log.info("총 요청 수: {}", CONCURRENT_USERS);
 		log.info("성공 요청 수: {}", successCount.get());
 		log.info("실패 요청 수: {}", failCount.get());
-		log.info("DB 기준 발급 수 (userCoupon): {}", issuedCount);
+		log.info("DB 기준 발급 수(userCoupon): {}", issuedCount);
 		log.info("남은 수량: {}", remainingQuantity);
 
 		Assertions.assertEquals(COUPON_QUANTITY, issuedCount, "정확한 수량만큼 발급돼야 함");
 		Assertions.assertEquals(successCount.get(), COUPON_QUANTITY, "성공 요청 수도 수량과 같아야 함");
-	}
-
-	@Test
-	@DisplayName("단일 사용자 중복 발급 방지 테스트")
-	void 중복_발급_방지_테스트() throws InterruptedException {
-		int threadCount = 10;
-		ExecutorService executorService = Executors.newFixedThreadPool(10);
-		CountDownLatch latch = new CountDownLatch(threadCount);
-
-		// 첫 번째 사용자로 중복 발급 테스트
-		User testUser = users.get(0);
-
-		for (int i = 0; i < threadCount; i++) {
-			final int idx = i;
-
-			executorService.submit(() -> {
-				try {
-					couponIssueFacade.issueCoupon(testUser.getId(), coupon.getId());
-					System.out.println("Thread " + idx + " - 발급 성공");
-
-				} catch (BusinessException e) {
-					System.out.println("Thread " + idx + " - 예외: " + e.getErrorCode());
-
-				} catch (Exception e) {
-					System.out.println("Thread " + idx + " - 에러: " + e.getMessage());
-
-				} finally {
-					latch.countDown();
-				}
-			});
-		}
-
-		latch.await();
-		executorService.shutdown();
-
-		Long issuedCount = userCouponRepository.countByUserIdAndCouponId(testUser.getId(), coupon.getId());
-
-		System.out.println("해당 사용자에게 발급된 쿠폰 수: " + issuedCount);
-
-		assertThat(issuedCount).isEqualTo(1L);
 	}
 }
