+++ src/test/java/com/threestar/trainus/domain/coupon/user/service/CouponServiceTest.java
@@ -1,39 +1,46 @@
 package com.threestar.trainus.domain.coupon.user.service;
 
 import static org.assertj.core.api.Assertions.*;
-import static org.mockito.ArgumentMatchers.*;
 import static org.mockito.BDDMockito.*;
 
 import java.time.LocalDateTime;
+import java.util.List;
 import java.util.Optional;
 
-import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.springframework.context.annotation.Import;
 
+import com.threestar.trainus.domain.coupon.user.dto.CouponPageResponseDto;
+import com.threestar.trainus.domain.coupon.user.dto.CouponResponseDto;
+import com.threestar.trainus.domain.coupon.user.dto.CreateUserCouponResponseDto;
+import com.threestar.trainus.domain.coupon.user.dto.UserCouponPageResponseDto;
 import com.threestar.trainus.domain.coupon.user.entity.Coupon;
 import com.threestar.trainus.domain.coupon.user.entity.CouponCategory;
 import com.threestar.trainus.domain.coupon.user.entity.CouponStatus;
 import com.threestar.trainus.domain.coupon.user.entity.UserCoupon;
 import com.threestar.trainus.domain.coupon.user.repository.CouponRepository;
 import com.threestar.trainus.domain.coupon.user.repository.UserCouponRepository;
+import com.threestar.trainus.domain.coupon.user.service.CouponService;
 import com.threestar.trainus.domain.user.entity.User;
 import com.threestar.trainus.domain.user.entity.UserRole;
 import com.threestar.trainus.domain.user.service.UserService;
+import com.threestar.trainus.global.config.JpaAuditingConfig;
 import com.threestar.trainus.global.exception.domain.ErrorCode;
 import com.threestar.trainus.global.exception.handler.BusinessException;
 
 @ExtendWith(MockitoExtension.class)
-class CouponServiceTest {
-
-	@Mock
-	private CouponRepository couponRepository;
+@Import(JpaAuditingConfig.class)
+public class CouponServiceTest {
 
 	@Mock
 	private UserCouponRepository userCouponRepository;
+
+	@Mock
+	private CouponRepository couponRepository;
 
 	@Mock
 	private UserService userService;
@@ -42,122 +49,231 @@ class CouponServiceTest {
 	private CouponService couponService;
 
 	@Test
-	@DisplayName("퍼센트 할인 쿠폰은 원가 기준 할인 금액을 계산한다")
-	void calculateDiscountedPrice_Percentage() {
-		UserCoupon userCoupon = new UserCoupon(createUser(1L), createCoupon("30%"), LocalDateTime.now().plusDays(10));
-
-		int discount = couponService.calculateDiscountedPrice(50000, userCoupon);
-
-		assertThat(discount).isEqualTo(15000);
-	}
-
-	@Test
-	@DisplayName("정액 할인 쿠폰은 원 단위 문자열에서 할인 금액을 계산한다")
-	void calculateDiscountedPrice_FixedAmount() {
-		UserCoupon userCoupon = new UserCoupon(createUser(1L), createCoupon("5000원"), LocalDateTime.now().plusDays(10));
-
-		int discount = couponService.calculateDiscountedPrice(50000, userCoupon);
-
-		assertThat(discount).isEqualTo(5000);
-	}
-
-	@Test
-	@DisplayName("사용자와 쿠폰 ID로 조회한 ACTIVE 쿠폰만 유효 쿠폰으로 반환한다")
-	void getValidUserCoupon_Active() {
+	void createUserCouponWithDistributedLock_정상_발급() {
 		Long userId = 1L;
-		Long couponId = 10L;
-		UserCoupon userCoupon = new UserCoupon(createUser(userId), createCoupon("5000원"), LocalDateTime.now().plusDays(10));
-		given(userCouponRepository.findByUserIdAndCouponId(userId, couponId)).willReturn(Optional.of(userCoupon));
+		Long couponId = 2L;
 
-		UserCoupon result = couponService.getValidUserCoupon(couponId, userId);
+		User user = createMockUser();
 
-		assertThat(result).isSameAs(userCoupon);
-		verify(userCouponRepository).findByUserIdAndCouponId(userId, couponId);
+		Coupon coupon = Coupon.builder()
+			.id(couponId)
+			.category(CouponCategory.NORMAL)
+			.closeAt(LocalDateTime.now().plusDays(1))
+			.expirationDate(LocalDateTime.now().plusDays(30))
+			.openAt(LocalDateTime.now())
+			.build();
+
+		given(userService.getUserById(userId))
+			.willReturn(user);
+		given(couponRepository.findById(couponId))
+			.willReturn(Optional.of(coupon));
+		given(userCouponRepository.existsByUserIdAndCouponId(userId, couponId))
+			.willReturn(false);
+
+		CreateUserCouponResponseDto responseDto = couponService.issueCoupon(userId, couponId);
+
+		assertThat(responseDto).isNotNull();
+		assertThat(responseDto.couponId()).isEqualTo(couponId);
+		then(userCouponRepository).should().save(any(UserCoupon.class));
+
 	}
 
 	@Test
-	@DisplayName("사용자와 쿠폰 ID로 조회한 INACTIVE 쿠폰은 유효 쿠폰 조회시 예외 처리한다")
-	void getValidUserCoupon_InactiveThrows() {
+	void createUserCouponWithDistributedLock_종료시각_지나면_예외처리() {
 		Long userId = 1L;
-		Long couponId = 10L;
-		UserCoupon userCoupon = new UserCoupon(createUser(userId), createCoupon("5000원"), LocalDateTime.now().plusDays(10));
-		userCoupon.use();
-		given(userCouponRepository.findByUserIdAndCouponId(userId, couponId)).willReturn(Optional.of(userCoupon));
+		Long couponId = 2L;
 
-		assertThatThrownBy(() -> couponService.getValidUserCoupon(couponId, userId))
+		User user = createMockUser();
+
+		Coupon coupon = Coupon.builder()
+			.id(couponId)
+			.category(CouponCategory.NORMAL)
+			.closeAt(LocalDateTime.now().minusDays(1))
+			.expirationDate(LocalDateTime.now().plusDays(30))
+			.build();
+
+		given(userService.getUserById(userId))
+			.willReturn(user);
+		given(couponRepository.findById(couponId))
+			.willReturn(Optional.of(coupon));
+
+		assertThatThrownBy(() -> couponService.issueCoupon(userId, couponId))
 			.isInstanceOf(BusinessException.class)
-			.hasFieldOrPropertyWithValue("errorCode", ErrorCode.COUPON_NOT_FOUND);
-		verify(userCouponRepository).findByUserIdAndCouponId(userId, couponId);
+			.hasMessageContaining(ErrorCode.COUPON_EXPIRED.getMessage());
 	}
 
 	@Test
-	@DisplayName("ACTIVE 쿠폰 사용시 상태를 INACTIVE로 바꾸고 저장한다")
-	void useCoupon_ActiveSaves() {
-		UserCoupon userCoupon = new UserCoupon(createUser(1L), createCoupon("5000원"), LocalDateTime.now().plusDays(10));
+	void createUserCouponWithDistributedLock_중복발급_예외처리() {
+		Long userId = 1L;
+		Long couponId = 2L;
 
-		couponService.useCoupon(userCoupon);
+		User user = createMockUser();
 
-		assertThat(userCoupon.getStatus()).isEqualTo(CouponStatus.INACTIVE);
-		verify(userCouponRepository).save(userCoupon);
+		Coupon coupon = Coupon.builder()
+			.id(couponId)
+			.category(CouponCategory.NORMAL)
+			.closeAt(LocalDateTime.now().plusDays(1))
+			.expirationDate(LocalDateTime.now().plusDays(30))
+			.build();
+
+		given(userService.getUserById(userId))
+			.willReturn(user);
+		given(couponRepository.findById(couponId))
+			.willReturn(Optional.of(coupon));
+		given(userCouponRepository.existsByUserIdAndCouponId(userId, couponId))
+			.willReturn(true);
+
+		assertThatThrownBy(() -> couponService.issueCoupon(userId, couponId))
+			.isInstanceOf(BusinessException.class)
+			.hasMessageContaining(ErrorCode.COUPON_ALREADY_ISSUED.getMessage());
+
 	}
 
 	@Test
-	@DisplayName("INACTIVE 쿠폰 사용 요청은 저장하지 않는다")
-	void useCoupon_InactiveDoesNothing() {
-		UserCoupon userCoupon = new UserCoupon(createUser(1L), createCoupon("5000원"), LocalDateTime.now().plusDays(10));
-		userCoupon.use();
+	void createUserCouponWithDistributedLock_선착순쿠폰_오픈전_예외처리() {
+		Long userId = 1L;
+		Long couponId = 2L;
 
-		couponService.useCoupon(userCoupon);
+		User user = createMockUser();
 
-		assertThat(userCoupon.getStatus()).isEqualTo(CouponStatus.INACTIVE);
-		verify(userCouponRepository, never()).save(any(UserCoupon.class));
+		Coupon coupon = Coupon.builder()
+			.id(couponId)
+			.openAt(LocalDateTime.now().plusDays(1))
+			.category(CouponCategory.OPEN_RUN)
+			.closeAt(LocalDateTime.now().plusDays(1))
+			.expirationDate(LocalDateTime.now().plusDays(30))
+			.build();
+
+		given(userService.getUserById(userId))
+			.willReturn(user);
+		given(couponRepository.findById(couponId))
+			.willReturn(Optional.of(coupon));
+
+		assertThatThrownBy(() -> couponService.issueCoupon(userId, couponId))
+			.isInstanceOf(BusinessException.class)
+			.hasMessageContaining(ErrorCode.COUPON_NOT_YET_OPEN.getMessage());
 	}
 
 	@Test
-	@DisplayName("INACTIVE 쿠폰 복원시 ACTIVE로 바꾸고 저장한다")
-	void restoreCoupon_InactiveSaves() {
-		UserCoupon userCoupon = new UserCoupon(createUser(1L), createCoupon("5000원"), LocalDateTime.now().plusDays(10));
-		userCoupon.use();
+	void createUserCouponWithDistributedLock_수량소진_예외() {
+		Long userId = 1L;
+		Long couponId = 2L;
 
-		couponService.restoreCoupon(userCoupon);
+		User user = createMockUser();
 
-		assertThat(userCoupon.getStatus()).isEqualTo(CouponStatus.ACTIVE);
-		assertThat(userCoupon.getUseDate()).isNull();
-		verify(userCouponRepository).save(userCoupon);
+		Coupon coupon = Coupon.builder()
+			.id(couponId)
+			.quantity(0)
+			.openAt(LocalDateTime.now().minusDays(1))
+			.category(CouponCategory.OPEN_RUN)
+			.closeAt(LocalDateTime.now().plusDays(1))
+			.expirationDate(LocalDateTime.now().plusDays(30))
+			.build();
+
+		given(userService.getUserById(userId))
+			.willReturn(user);
+		given(couponRepository.findById(couponId))
+			.willReturn(Optional.of(coupon));
+		assertThatThrownBy(() -> couponService.issueCoupon(userId, couponId))
+			.isInstanceOf(BusinessException.class)
+			.hasMessageContaining(ErrorCode.COUPON_BE_EXHAUSTED.getMessage());
 	}
 
 	@Test
-	@DisplayName("ACTIVE 쿠폰 복원 요청은 저장하지 않는다")
-	void restoreCoupon_ActiveDoesNothing() {
-		UserCoupon userCoupon = new UserCoupon(createUser(1L), createCoupon("5000원"), LocalDateTime.now().plusDays(10));
+	void getUserCoupons_상태없는_성공케이스() {
+		Long userId = 1L;
+		User user = createMockUser();
 
-		couponService.restoreCoupon(userCoupon);
-
-		assertThat(userCoupon.getStatus()).isEqualTo(CouponStatus.ACTIVE);
-		verify(userCouponRepository, never()).save(any(UserCoupon.class));
-	}
-
-	private Coupon createCoupon(String discountPrice) {
-		LocalDateTime now = LocalDateTime.now();
-		return Coupon.builder()
-			.name("테스트 쿠폰")
-			.discountPrice(discountPrice)
+		Coupon coupon = Coupon.builder()
+			.id(1L)
+			.name("10% 할인 쿠폰")
+			.discountPrice("10%")
 			.minOrderPrice(10000)
-			.status(CouponStatus.ACTIVE)
+			.expirationDate(LocalDateTime.now().plusDays(7))
 			.quantity(100)
 			.category(CouponCategory.NORMAL)
-			.openAt(now.minusDays(1))
-			.closeAt(now.plusDays(1))
-			.expirationDate(now.plusDays(30))
+			.openAt(LocalDateTime.now().minusDays(1))
+			.closeAt(LocalDateTime.now().plusDays(3))
 			.build();
+
+		UserCoupon userCoupon = new UserCoupon(user, coupon, coupon.getExpirationDate());
+
+		willDoNothing().given(userService).validateUserExists(userId);
+
+		given(userCouponRepository.findAllByUserIdWithCoupon(userId))
+			.willReturn(List.of(userCoupon));
+
+		UserCouponPageResponseDto dto = couponService.getUserCoupons(userId, null);
+
+		assertThat(dto).isNotNull();
+		assertThat(dto.userCoupons()).hasSize(1);
 	}
 
-	private User createUser(Long userId) {
+	@Test
+	void getUserCoupons_상태있는_성공케이스() {
+		Long userId = 1L;
+		CouponStatus status = CouponStatus.INACTIVE;
+		User user = createMockUser();
+
+		Coupon coupon = Coupon.builder()
+			.id(1L)
+			.name("10% 할인 쿠폰")
+			.discountPrice("10%")
+			.minOrderPrice(10000)
+			.expirationDate(LocalDateTime.now().plusDays(7))
+			.quantity(100)
+			.category(CouponCategory.NORMAL)
+			.openAt(LocalDateTime.now().minusDays(1))
+			.closeAt(LocalDateTime.now().plusDays(3))
+			.status(CouponStatus.INACTIVE)
+			.build();
+
+		Coupon coupon2 = Coupon.builder()
+			.id(2L)
+			.name("10% 할인 쿠폰")
+			.discountPrice("10%")
+			.minOrderPrice(10000)
+			.expirationDate(LocalDateTime.now().plusDays(7))
+			.quantity(100)
+			.category(CouponCategory.NORMAL)
+			.openAt(LocalDateTime.now().minusDays(1))
+			.closeAt(LocalDateTime.now().plusDays(3))
+			.status(CouponStatus.ACTIVE)
+			.build();
+
+		UserCoupon userCoupon = new UserCoupon(user, coupon, coupon.getExpirationDate());
+		willDoNothing().given(userService).validateUserExists(userId);
+		given(userCouponRepository.findAllByUserIdAndStatusWithCoupon(userId, status))
+			.willReturn(List.of(userCoupon));
+
+		UserCouponPageResponseDto dto = couponService.getUserCoupons(userId, status);
+
+		assertThat(dto).isNotNull();
+		assertThat(dto.userCoupons()).hasSize(1);
+		assertThat(dto.userCoupons().get(0).couponId()).isEqualTo(coupon.getId());
+	}
+
+	@Test
+	void getCoupons_성공_테스트() {
+		Long userId = 1L;
+
+		List<CouponResponseDto> dummyDtoList = List.of(mock(CouponResponseDto.class));
+		willDoNothing().given(userService).validateUserExists(userId);
+		given(couponRepository.findAvailableCouponsWithOwnership(eq(userId), any())).willReturn(dummyDtoList);
+
+		// when
+		CouponPageResponseDto result = couponService.getCoupons(userId);
+
+		// then
+		assertThat(result).isNotNull();
+		assertThat(result.coupons()).hasSize(1);
+	}
+
+	private User createMockUser() {
 		return User.builder()
-			.id(userId)
-			.email("test" + userId + "@test.com")
-			.password("encoded")
-			.nickname("테스트유저" + userId)
+			.email("user@test.com")
+			.password("test1234")
+			.nickname("user")
 			.role(UserRole.USER)
 			.build();
 	}
